CTCP-282: Added test shell scripts and runners
