+++ drop_unloading_frontend_data.sh
@@ -1,0 +1,6 @@
+#!/usr/bin/env bash
+{
+  sleep 1
+  echo "use manage-transit-movements-unloading-frontend"
+  echo "db['user-answers'].drop()"
+} | mongo
