+++ run_a11y_tests.sh
@@ -1,0 +1,2 @@
+#!/usr/bin/env bash
+sbt -Dbrowser=remote-chrome -Da11y=true -Denvironment=local 'testOnly uk.gov.hmrc.test.ui.cucumber.runner.RunnerA11y'
