+++ src/test/resources/features/TraderDetails.feature
@@ -9,6 +9,7 @@ Feature: Trader details - includes Pre-task list, Holder of transit, Representat
     Given I click on the Make a departure declaration link
     And I input a random LRN on the LocalReferenceNumber page
 
+  @a11y @ZAP
   Scenario: 01 TraderDetails - GB Normal - No Security - Reduced Data Set - Not more than 1 consignee
     And I select GB on the office of departure page
     And I choose radio option Normal on the procedure type page
@@ -223,14 +224,3 @@ Feature: Trader details - includes Pre-task list, Holder of transit, Representat
     And I submit on the CYA page
     Then I should be on the task list page
     And I should see COMPLETED status for trader details
-
-
-
-
-
-
-
-
-
-
-
