+++ src/test/scala/uk/gov/hmrc/test/ui/cucumber/runner/RunnerForManage.scala
@@ -14,23 +14,16 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.test.ui.pages.TraderDetails.Guarantee
+package uk.gov.hmrc.test.ui.cucumber.runner
 
-import uk.gov.hmrc.test.ui.pages.BasePage
+import io.cucumber.junit.{Cucumber, CucumberOptions}
+import org.junit.runner.RunWith
 
-object GuaranteeReferenceNumberPage extends BasePage {
-
-  val guaranteeReferenceNumberTitle =
-    "What is the Guarantee Reference Number (GRN)? - Manage your transit movements - GOV.UK"
-
-  def loadPage: this.type = {
-    onPage(guaranteeReferenceNumberTitle)
-    this
-  }
-
-  def enterGRN(guaranteeReferenceNumber: String): GuaranteeReferenceNumberPage.type = {
-    fillInputById("value", guaranteeReferenceNumber)
-    this
-  }
-
-}
+@RunWith(classOf[Cucumber])
+@CucumberOptions(
+  features = Array("src/test/resources/features"),
+  glue = Array("uk.gov.hmrc.test.ui.cucumber.stepdefs"),
+  plugin = Array("pretty", "html:target/cucumber", "json:target/cucumber.json"),
+  tags = "@manage_transit_movements or @managelegacyenrolment"
+)
+class RunnerForManage {}
