+++ src/test/scala/uk/gov/hmrc/test/ui/cucumber/runner/RunnerWip.scala
@@ -14,23 +14,16 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.test.ui.pages.TraderDetails.Guarantee
+package uk.gov.hmrc.test.ui.cucumber.runner
 
-import uk.gov.hmrc.test.ui.pages.BasePage
+import io.cucumber.junit.{Cucumber, CucumberOptions}
+import org.junit.runner.RunWith
 
-object OtherGuarantee3ReferencePage extends BasePage {
-
-  val otherGuarantee3ReferenceTitle =
-    "What is the reference for the guarantee? - Manage your transit movements - GOV.UK"
-
-  def loadPage: this.type = {
-    onPage(otherGuarantee3ReferenceTitle)
-    this
-  }
-
-  def enterOtherReference(answer: String): OtherGuarantee3ReferencePage.type = {
-    fillInputById("value", answer)
-    this
-  }
-
-}
+@RunWith(classOf[Cucumber])
+@CucumberOptions(
+  features = Array("src/test/resources/features"),
+  glue = Array("uk.gov.hmrc.test.ui.cucumber.stepdefs"),
+  plugin = Array("pretty", "html:target/cucumber", "json:target/cucumber.json"),
+  tags = "@wip"
+)
+class RunnerWip {}
