+++ src/test/scala/uk/gov/hmrc/test/ui/pages/Guarantee/AddAnotherGuaranteePage.scala
@@ -1,0 +1,43 @@
+/*
+ * Copyright 2022 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package uk.gov.hmrc.test.ui.pages.Guarantee
+
+import uk.gov.hmrc.test.ui.pages.BasePage
+
+object AddAnotherGuaranteePage extends BasePage {
+
+  var numberOfGuarantees = 1
+  var guarantees         = "guarantee"
+
+  def addAnotherGuaranteeTitle =
+    f"You have added $numberOfGuarantees $guarantees - Manage your transit movements - GOV.UK"
+
+  def loadPage: this.type = {
+    onPage(addAnotherGuaranteeTitle)
+    this
+  }
+
+  def needToAddAnotherGuarantee(answer: String): Unit =
+    answer match {
+      case "Yes" =>
+        clickById("value")
+        numberOfGuarantees += 1
+        guarantees = "guarantees"
+      case "No"  => clickById("value-no");
+    }
+
+}
